refactor(api): move EmailModificationDemand model
to src
